Fixed ProductController (now only owner can make changes). ReviewDecorator works properly. Each product and review is now assign to creator.

+++ app/controllers/products_controller.rb
@@ -1,4 +1,6 @@
 class ProductsController < ApplicationController
+
+  before_action :author!, only: [:edit, :update]
   expose(:category)
   expose(:products)
   expose(:product)
@@ -18,33 +20,52 @@ class ProductsController < ApplicationController
   end
 
   def create
-    self.product = Product.new(product_params)
-
-    if product.save
-      category.products << product
-      redirect_to category_product_url(category, product), notice: 'Product was successfully created.'
+    if user_signed_in?
+      self.product = Product.new(product_params)
+      product.user_id = current_user.id
+      if product.save
+        category.products << product
+        redirect_to category_product_url(category, product), notice: 'Product was successfully created.'
+      else
+        render action: 'new'
+      end
     else
-      render action: 'new'
+      redirect_to new_user_session_path, notice: "You need to sign in first."
     end
   end
 
   def update
-    if self.product.update(product_params)
-      redirect_to category_product_url(category, product), notice: 'Product was successfully updated.'
+    if user_signed_in?
+      if self.product.update(product_params)
+        redirect_to category_product_url(category, product), notice: 'Product was successfully updated.'
+      else
+        render action: 'edit'
+      end
     else
-      render action: 'edit'
+      redirect_to new_user_session_path, notice: "You need to sign in first."
     end
   end
 
   # DELETE /products/1
   def destroy
-    product.destroy
-    redirect_to category_url(product.category), notice: 'Product was successfully destroyed.'
+    if user_signed_in?
+      product.destroy
+      redirect_to category_url(product.category), notice: 'Product was successfully destroyed.'
+    else
+      redirect_to new_user_session_path, notice: "You need to sign in first."
+    end
   end
 
   private
 
-  def product_params
-    params.require(:product).permit(:title, :description, :price, :category_id)
-  end
+    def product_params
+      params.require(:product).permit(:title, :description, :price, :category_id)
+    end
+
+    def author!
+      unless self.product.user == current_user
+        redirect_to category_product_url(category, product),
+          flash: { error: 'You are not allowed to edit this product.' }
+    end
+end
 end
